+++ RELEASE-NOTES.md
@@ -11,6 +11,27 @@ VERSION SCHEMA: vYYYY.MM.#SEQ
 
 ---
 
+## v2019.05.3
+
+- NEW: Add Delete-Criminal-Fingerprints workflow.
+- CHN: Allow skipping 9 fingerprints in criminal workflows without role.
+
+---
+
+## v2019.05.2
+
+- NEW: Add Register-Criminal-Fingerprints-For-Present-Person workflow.
+- NEW: Add Register-Criminal-Fingerprints-For-Not-Present-Person workflow.
+
+---
+
+## v2019.05.1
+
+- FIX: Using civil biometrics ID and criminal biometrics ID in register-convicted-report-for-non-present workflow is not complete.
+- FIX: Spaces are allowed in text fields and are not trimmed.
+
+---
+
 ## v2019.04.3
 
 - FIX: Disable update-personId-info on Convicted-Report-Registration on going back.